% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dygraph_functions.R
\name{e4_timeseries_plot}
\alias{e4_timeseries_plot}
\title{e4_timeseries_plot}
\usage{
e4_timeseries_plot(data, main_title = "", calendar_data = NULL,
  average_lines = "", yaxis_ranges = NULL, colours = c("#1874CD",
  "#FF7F00", "#458B00", "#68228B"))
}
\arguments{
\item{data}{List of e4 data.}

\item{main_title}{Main title of the plot, defaults to empty string.}

\item{calendar_data}{Name to give to the timeseries data.}

\item{average_lines}{Do you want to add averages to the plot}

\item{yaxis_ranges}{What Range do you want to give the y axis?}

\item{colours}{The colours of the graph, defaults to blue, orange, red, green}
}
\description{
Creates 4 graphs of the E4 data.
}
