% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e4server.R
\name{e4server}
\alias{e4server}
\title{e4server}
\usage{
e4server(input, output, session)
}
\arguments{
\item{input}{standard input, output, session follwing Shinyapp}

\item{output}{standard input, output, session follwing Shinyapp}

\item{session}{standard input, output, session follwing Shinyapp}
}
\description{
responsible for server transformations
}
